perf(so3ssd): carry the chunk-start state at the activation dtype
zstart and dinc are (B,H,C,P,3N) planes no recurrence reads back: the forward
writes zstart out of registers, and every consumer stages it through shared
memory, where table.stage_state already narrowed it to the operand dtype.
Float32 storage paid for a width nothing used.

Both now take the dtype of U. Invariant 4 pins float32 to the recurrence that
produces a state, not to the copy a later GEMM reads, so the pinned tuple in
five kernels splits: the pinned operands keep check_pinned, and the two stored
buffers take check_stored against the activation dtype.

Acceptance shape, both arms in one process on one A6000. Durations are profiler
medians of nine steps at 1.76 to 1.80 GHz; traffic is NCU medians of three
launches at 1.39 to 1.41 GHz:

  increment_passing_fwd    397.2 ->  370.1 us   write 143.67 ->  73.46 MB
  chunk_scan_fwd           573.6 ->  498.3 us   read  176.19 -> 104.71 MB
  start_passing_bwd        409.4 ->  348.8 us   write 143.72 ->  73.46 MB
  chunk_input_bwd          881.0 ->  761.9 us   read  484.01 -> 259.25 MB
  chunk_vector_bwd       1,878.4 -> 1,863.1 us  read  336.64 -> 195.03 MB

-297.5 us of kernel time for -609.7 MB of DRAM traffic, -261 us on the step,
and -272.00 MiB of peak allocation, of which -68.00 MiB is the retained forward
state by itself. chunk_scan_fwd also spills less: 1,032,192 fewer local sectors
each way. Deleted bytes convert at 0.334 of face value overall and 0.266 on the
fused prologue, which is already off the bus at 65.4 percent of memory SoL.
chunk_vector_bwd is the exception and reads as flat: -15.5 us on the profiler
against +15.5 us under NCU, on a 0.03 percent shift in active cycles.

The forward is bit-exact, since stage_state already narrowed the value that now
reaches memory: 0 of 9,437,184 elements move. The backward moves elements but
not its envelope: against float64 through the reference, dB measures 4.637e-03
against its 1.0e-02 bound, dU 4.932e-03 against 1.1e-02, dK 2.745e-03 against
7.0e-03, dtrans 6.294e-03 against 2.0e-02, and dC 6.057e-03 against 1.0e-02,
each equal to the float32-buffer arm's figure at four significant digits. The
maxima sit on the operand cast, which both arms share; dC does not move at all.
Inside the chunk-input stage the pair's rounding does reach dchunk_rot, whose
oracle is not evaluated at the pair: 1.3x in bfloat16 and 1.9x in float16, to
77 and 85 percent of a bound that does not move.

float16 also holds, with one hazard recorded in the module docstring: dinc's
smallest nonzero at this shape is 2.537e-09, and 9.325e-07 of its elements sit
below float16's smallest subnormal and flush to zero.

+++ scripts/perf/profile_chunk_input_bwd.py
@@ -189,8 +189,8 @@ def build_runner(
     chunks = -(-shape.seq // shape.chunk)
     gen = torch.Generator(device=device).manual_seed(1)
     state = (shape.bsz, shape.heads, chunks, shape.rows, shape.d_state)
-    dinc = torch.randn(*state, dtype=torch.float32, device=device, generator=gen)
-    zstart = torch.randn(*state, dtype=torch.float32, device=device, generator=gen)
+    dinc = torch.randn(*state, dtype=dtype, device=device, generator=gen)
+    zstart = torch.randn(*state, dtype=dtype, device=device, generator=gen)
     forcing = inputs.B[:, :groups]
     readout = inputs.C[:, :groups]
